nomsg

+++ src/lib/HTTP/HTTPServer.cpp
@@ -595,10 +595,11 @@ bool SendResponse(CHTTPSessionInfo* p_Session, CHTTPMessage* p_Response, CHTTPMe
   
   
   // send binary
-  unsigned int nOffset      = 0;
-  char*        szChunk      = NULL;
-  unsigned int nRequestSize = MAX_BUFFER_SIZE;
-  int          nErr         = 0;    
+  //unsigned int nOffset      = 0;
+  off_t         nOffset       = 0;
+  char*         szChunk       = NULL;
+  unsigned int  nRequestSize  = MAX_BUFFER_SIZE;
+  int           nErr          = 0;    
     
     
   // calculate response ranges and request size
